+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.8.0] - 2022-06-08
+### Added
+- [#26] Allow the definition of custom Deployment in dogus. In such custom Deployments it is possible
+  to define extra volumes,volume mounts, and the used service account for the dogu Deployment.
+
 ## [v0.7.0] - 2022-06-07
 ### Added
 - [#6] Installing generic kubernetes resources when installing a dogu. These resources need to be provided by the dogu image